Remove Generality in info provider
